requirements.txt
alembic==1.15.2
APScheduler==3.11.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Faker==37.1.0
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.0.2
paho-mqtt==2.1.0
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
requests==2.32.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.13.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.40
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.21.0
